cleanup

+++ UniversityDeadlineTracker/ClientApp/src/Utils/Token.js
@@ -1,31 +1,28 @@
-﻿import {useState} from 'react';
+﻿import { useState } from "react";
 
 export const getToken = () => {
-    
-    return sessionStorage.getItem('token');
+    return sessionStorage.getItem("token");
 };
 
 export const getUser = () => {
-    return JSON.parse(sessionStorage.getItem('user'));
+    return JSON.parse(sessionStorage.getItem("user"));
 };
 
 export const useToken = () => {
-
     const [token, setToken] = useState(getToken());
 
     const saveToken = (userToken, refreshToken, userData) => {
-        if (userToken === null && userData === null)
-            sessionStorage.clear();
+        if (userToken === null && userData === null) sessionStorage.clear();
         else {
-            sessionStorage.setItem('token', userToken);
-            sessionStorage.setItem('refreshToken', refreshToken);
-            sessionStorage.setItem('user', JSON.stringify(userData));
+            sessionStorage.setItem("token", userToken);
+            sessionStorage.setItem("refreshToken", refreshToken);
+            sessionStorage.setItem("user", JSON.stringify(userData));
         }
         setToken(userToken);
     };
 
     return {
         token: token,
-        setToken: saveToken
-    }
-}
+        setToken: saveToken,
+    };
+};
